% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_instruments.R
\name{check_project_exists}
\alias{check_project_exists}
\title{check_project_exists}
\usage{
check_project_exists(redcap_uri, token)
}
\arguments{
\item{redcap_uri}{the REDCap URL}

\item{token}{the API key}
}
\value{
logical
}
\description{
This checks to make sure that the user specified a valid project
}
