% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_labels.R
\name{drop_labels}
\alias{drop_labels}
\title{Drop all the labels from a variable}
\usage{
drop_labels(df)
}
\arguments{
\item{df}{The data frame with column labels that you want to drop}
}
\value{
df without column labels
}
\description{
There is an issue with the function we are using to add column
labels.  If you run into problems processing the labels.
}
\examples{
\dontrun{
demographics |>
  drop_labels() |>
  skimr::skim()
}
}
