% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_variable.R
\name{find_variable}
\alias{find_variable}
\title{Find a variable in a dataset}
\usage{
find_variable(var_name)
}
\arguments{
\item{var_name}{Character or length one. The quoted variable to find in the Global
Environment.}
}
\description{
When using tidyREDCap projects, a variable may appear in multiple
datasets. This allows users to easily find which dataset(s) a variable name is in
from the Global Environment.
}
\examples{
find_variable('record_id')
}
