% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_choose_all_table.R
\name{getLabel2}
\alias{getLabel2}
\title{Get a Variable Label from a Data Set and Variable}
\usage{
getLabel2(data, aVariable)
}
\arguments{
\item{data}{The name of the data set}

\item{aVariable}{The name of the variable}
}
\value{
A variable's response label without  the choose all the question
}
\description{
Get a Variable Label from a Data Set and Variable
}
