% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_instruments.R
\name{import_instruments}
\alias{import_instruments}
\title{import_instruments}
\usage{
import_instruments(connection, drop_blank = TRUE, envir = .GlobalEnv)
}
\arguments{
\item{connection}{The name of connection created by a call to the
\code{redcapAPI::redcapConnection()} function}

\item{drop_blank}{Drop records that have no data. TRUE by default.}

\item{envir}{The name of the environment where the tables should be saved.}
}
\value{
datasets, by default in the global environment
}
\description{
This function takes the name of a connection to a REDCap
project and returns a table for each instrument/form in the project.
}
\examples{
\dontrun{
  rcon <- redcapAPI::redcapConnection(
    url <- "https://redcap.miami.edu/api/",
    token = Sys.getenv("REDCAP_PLACEHOLDER5_KEY")
  )
  
  import_instruments(rcon)
}   
}
