% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_choose_all_table.R
\name{make_choose_all_table}
\alias{make_choose_all_table}
\title{Count The Reponses to a Choose All That Apply Question}
\usage{
make_choose_all_table(df, variable)
}
\arguments{
\item{df}{The name of the data set}

\item{variable}{The name of the REDCap variable}
}
\value{
A variable's response label without  the choose all the question
}
\description{
Count The Reponses to a Choose All That Apply Question
}
