% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_choose_one_table.R
\name{make_choose_one_table}
\alias{make_choose_one_table}
\title{make_choose_one_table}
\usage{
make_choose_one_table(arg1, arg2, subset = FALSE)
}
\arguments{
\item{arg1}{data frame that has a factor or a factor name}

\item{arg2}{if arg1 is a data frame this is the factor name}

\item{subset}{TRUE/FALSE Remove extra variable name text from the label if used with a
choose all question}
}
\value{
a table
}
\description{
Pass this function either a labeled factor or a data frame and
a factor in the fame and it will return a janitor style table.  Use subset =
TRUE if you are making a report on varaible that is part of a choose all that
apply question.
}
