% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_instrument_auto.R
\name{make_instrument_auto}
\alias{make_instrument_auto}
\title{Extract an Instrument from an REDCap Export without specifying Variables}
\usage{
make_instrument_auto(df, drop_which_when = FALSE)
}
\arguments{
\item{df}{A data frame with the instrument}

\item{drop_which_when}{Drop the \code{record_id} and \code{redcap_event_name} variables}
}
\value{
A data frame that has an instrument (with at least one not NA value).
}
\description{
This function takes a data frame holding REDCap data,
checks if it is a longitudinal study, and returns records that have values.
}
