% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_yes_no.R
\name{make_yes_no}
\alias{make_yes_no}
\title{make_yes_no}
\usage{
make_yes_no(x)
}
\arguments{
\item{x}{x variable to be converted to hold "Yes" or "No or Unknown"}
}
\value{
a factor with "Yes" or "No or Unknown"
}
\description{
Convert a "Yes-No", "True-False", or "Checkboxes (Multiple
Answers)" question in REDCap to a factor holding "Yes" or
"No or Unknown". Technically "yes" or "checked" (ignoring case), 1 or
TRUE responses are converted to "Yes" and all other values to
"No or Unknown". Also see \code{make_yes_no_unknown()}.
}
\examples{
make_yes_no(c(0, 1, NA))
make_yes_no(c("unchecked", "Checked", NA))
}
