% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_choose_one_table.R
\name{taybull}
\alias{taybull}
\title{taybull function}
\usage{
taybull(variable, subset = FALSE)
}
\arguments{
\item{variable}{a factor to report on}

\item{subset}{parse off extra variable if used with a choose all question}
}
\value{
a table
}
\description{
Make a labeled janitor table from a single variable
}
