% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_choose_one_table.R
\name{taybull2}
\alias{taybull2}
\title{taybull2}
\usage{
taybull2(data, aVariable, subset = FALSE)
}
\arguments{
\item{data}{a dataframe that has the factor to report on}

\item{aVariable}{a factor to report on}

\item{subset}{TRUE/FALSE used to remove the repeated text from of a
\emph{choose all that apply} question}
}
\value{
a table
}
\description{
Make a labeled janitor table from a dataset and variable
}
