% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_instrument_auto.R
\name{drop_label}
\alias{drop_label}
\title{Drop the label from a variable}
\usage{
drop_label(df, x)
}
\arguments{
\item{df}{the name of the data frame}

\item{x}{the quoted name of the variable}
}
\value{
df
}
\description{
There is a reported issues with joins on data (without a reprex)
that seem to be caused by the labels.  As a possible solution this can be
used to drop labels.
}
