% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_binary_word.R
\name{make_binary_word}
\alias{make_binary_word}
\title{Convert a "choose all that apply" Question Into a Binary Word}
\usage{
make_binary_word(df, yes_value = "Checked", the_labels = letters)
}
\arguments{
\item{df}{A dataframe with the variables corresponding to binary indicators
(the dummy coded variables) for a "choose all that apply" question.}

\item{yes_value}{A character string that corresponds to choosing "yes"
in the df binary variables.  Defaults to the REDCap "Checked" option.}

\item{the_labels}{A character vector of single letters holding the letters
used to make the binary word.  See the article/vignette called "Make
Binary Word" for an example.}
}
\value{
A character vector with length equal to the rows of \code{df}
}
\description{
This function takes a dataframe holding binary variables with
values corresponding to a dummy-coded "choose all that apply" question.
It can be used for any \emph{binary word} problem.
}
\examples{
  test_df <- tibble::tibble(
    q1 = c("Unchecked", "Checked"),
    q2 = c("Unchecked", "Unchecked"),
    q3 = c("Checked", "Checked"),
    q4 = c("Checked", "Unchecked")
  )
  make_binary_word(test_df)
  
}
