% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_binary_word.R
\name{make_binary_word}
\alias{make_binary_word}
\title{Convert a choose all \strong{question} into a binary word}
\usage{
make_binary_word(df, yes_value = "Checked", the_labels = letters)
}
\arguments{
\item{df}{the dataframe with the columns}

\item{yes_value}{what was checked}

\item{the_labels}{character vector for indicator letters}
}
\value{
character vector of length of rows of \code{df}
}
\description{
Pass in a set variable
}
