% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_choose_all_table.R
\name{make_choose_all_table}
\alias{make_choose_all_table}
\title{Count The Responses to a Choose All That Apply Question}
\usage{
make_choose_all_table(df, variable)
}
\arguments{
\item{df}{The name of the data set (it needs labels)}

\item{variable}{The name of the REDCap variable}
}
\value{
A variable's response label without  the choose all the question
}
\description{
This will tally the number of responses on a choose all that
apply question.  This function extracts the option name from the variable
labels.  So the data set needs to be labeled.  See
the \href{../doc/makeChooseAllTable.html}{Make a 'Choose All' Table} vignette
for help.
}
