% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_instruments_redcapr.R
\name{relabel}
\alias{relabel}
\title{relabel}
\usage{
relabel(x)
}
\arguments{
\item{x}{str variable label to be check for possible labels that need fixing}
}
\value{
vector text changed
}
\description{
Function to change labels to match REDCapAPI.  REDCapR labels
choose all that apply variables with "(choice= thingy)" vs. ": thingy" in
REDCapAPI
}
\examples{
\dontrun{
relabel("What ingredients do you currently crave? (choice=Chips)")
}
}
